Update plan management to display commercial data and fix data types
Update DTOs, API contracts, and backend logic to correctly handle plan limits as integers and display commercial data in the plan list.

+++ boioot/apps/backend/src/Boioot.Application/Features/Plans/DTOs/PlanLimitItem.cs
@@ -6,6 +6,6 @@ public class PlanLimitItem
     public string Key               { get; set; } = string.Empty;
     public string Name              { get; set; } = string.Empty;
     public string? Unit             { get; set; }
-    /// <summary>-1 = unlimited.</summary>
-    public decimal Value            { get; set; }
+    /// <summary>-1 = unlimited. Always an integer.</summary>
+    public int Value                { get; set; }
 }
